Moved non-tests to tests/helpers.py and remaining fixture to conftest.py
